Move chatbox to frontend core
