+++ packages/frontend-core/src/components/index.js
@@ -10,3 +10,4 @@ export { default as FilterUsers } from "./FilterUsers.svelte"
 export { default as ChangePasswordModal } from "./ChangePasswordModal.svelte"
 export { default as ProfileModal } from "./ProfileModal.svelte"
 export { default as PasswordRepeatInput } from "./PasswordRepeatInput.svelte"
+export { default as Chatbox } from "./Chatbox/index.svelte"
