+++ packages/frontend-core/src/icons/BBAI.svelte
